Add for loops
For loops are probably the most well tested (relatively speaking)
of all other constructs, but I don't fully get why they actually work.

The point of confusion is that when the initializer is a variable
declaration, executing the initializer must introduce the initializer
variable to the loop's parent scope because we don't enter the loop's
scope until we start executing its body. So, what happens when the
initializer shadows an outer scope variable? If the loop modifies the
initializer variable while executing its body, the side effect must
spill in outer scopes. This is undesired, of course, because when the
loop exits, we would want all the side effects from the loop-local
variables to go away. AND IT DOES, as per the current implementation. So
on the surface, the implementation of for loops seems correct, but I
don't know why.