Traversing the scope chain for function calls is taboo
When you call a function, a new scope is pushed to the interpreter. All
the nested scopes of statements such as if, while, switch etc. are
pushed on top of it. We must allow traversing the scope chain for nested
scopes (an `if` block may query names from its parent scope), but this
isn't true for function calls. When function A calls, function B, we
push B's scope and now treat it as a scope nested within the scope of A.
That's clearly wrong. We'll need to fix it at some point.